Rewrote ZendeskButton to lazy-load all resources related to Zendesk. Also converted to ts
